update graphic11 dbglinestrip, config bugfix

+++ Common/math/vector2_comm.h
@@ -74,7 +74,7 @@ namespace common
 
 	template<class T>
 	float Vector2Comm<T>::Length() const {
-		return sqrt(x*x + y * y);
+		return sqrt((float)(x*x + y*y));
 	}
 
 	template<class T>
